requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.7
django-celery-beat==2.2.1
django-common-helpers==0.9.2
django-environ==0.8.1
django-filter==21.1
django-modeltranslation==0.17.3
django-rest-multiple-models==2.1.3
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
environ==1.0
flower==1.0.0
gunicorn==20.1.0
httpie==2.6.0
humanize==3.12.0
idna==3.2
ipython==7.28.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-gettext==4.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-toolbelt==0.9.1
schedule==1.1.0
six==1.16.0
slack-bolt==1.9.2
slack-sdk==3.11.2
sqlparse==0.4.2
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
